Do not rely on the user_profile in directory-sso
user_profile may not necessarily be available in directory-sso so trying to
get user name from it is not reliable. It also has few other issues wherein
if the user updates the name in directory-sso then it results in inconsistency.
For now there is no need to save user name so ignore them completely.